refactor: entity class를 sql.entity로 이동
