+++ src/main/java/persistence/sql/entity/NewPerson.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.GeneratedValue;
